requirements.txt
numpy~=1.19.5
matplotlib~=3.3.4
distinctipy~=1.1.5
opencv-python~=4.5.1.48
pillow~=8.2.0
scikit-learn~=0.24.1
pandas~=1.1.5
tensorflow~=2.4.1
kneed~=0.7.0
six~=1.15.0
svglib
reportlab
flask_uploads
flask_wtf